Making the pop-up work
