+++ .storybook/stories/TableFiltering/ExampleMedium.tsx
@@ -70,81 +70,78 @@ export const ExampleMedium = ({
 	const bodyId = 'filter-body';
 
 	return (
-		<PageContent>
-			<Stack gap={2}>
-				<Prose>
-					<h1>Table Filtering (Medium)</h1>
-					<p>
-						The medium filtering pattern is for cases where the number of
-						filterable fields is between 1 and 4 secondary filters. Our
-						FilterBar is used to reveal all filterable fields when the button is
-						pressed.
-					</p>
-				</Prose>
-				<div>
-					<ButtonLink href="#new" iconBefore={PlusIcon}>
-						New item
-					</ButtonLink>
-				</div>
+		<Stack gap={2}>
+			<Prose>
+				<h1>Table Filtering (Medium)</h1>
+				<p>
+					The medium filtering pattern is for cases where the number of
+					filterable fields is between 1 and 4 secondary filters. Our FilterBar
+					is used to reveal all filterable fields when the button is pressed.
+				</p>
+			</Prose>
+			<div>
+				<ButtonLink href="#new" iconBefore={PlusIcon}>
+					New item
+				</ButtonLink>
+			</div>
 
-				<FilterRegion>
-					<FilterBar>
-						<FilterBarGroup>
-							<FilterSearchInput filters={filters} setFilters={setFilters} />
-							<FilterStatusSelect filters={filters} setFilters={setFilters} />
-							<Button
-								onClick={toggleIsOpen}
-								variant="secondary"
-								iconBefore={FilterIcon}
-								iconAfter={isOpen ? ChevronUpIcon : ChevronDownIcon}
-								// accessibility
-								aria-label="more filters"
-								id={buttonId}
-								aria-controls={bodyId}
-								aria-expanded={isOpen}
-							>
-								{isOpen ? 'Hide filters' : 'Show filters'}
-							</Button>
-						</FilterBarGroup>
+			<FilterRegion>
+				<FilterBar>
+					<FilterBarGroup>
+						<FilterSearchInput filters={filters} setFilters={setFilters} />
+						<FilterStatusSelect filters={filters} setFilters={setFilters} />
+						<Button
+							onClick={toggleIsOpen}
+							variant="secondary"
+							iconBefore={FilterIcon}
+							iconAfter={isOpen ? ChevronUpIcon : ChevronDownIcon}
+							// accessibility
+							aria-label="more filters"
+							id={buttonId}
+							aria-controls={bodyId}
+							aria-expanded={isOpen}
+						>
+							{isOpen ? 'Hide filters' : 'Show filters'}
+						</Button>
+					</FilterBarGroup>
 
-						<SortBySelect sort={sort} setSort={setSort} />
-					</FilterBar>
-					<FilterAccordion
-						id={bodyId}
-						ariaLabelledBy={buttonId}
-						filters={filters}
-						isOpen={isOpen}
-						setFilters={setFilters}
-						resetFilters={resetFilters}
-					/>
-					<ActiveFilters
-						filters={filters}
-						removeFilter={removeFilter}
-						resetFilters={resetFilters}
-					/>
-				</FilterRegion>
-
-				<DashboardTable
-					data={data}
-					loading={loading}
-					itemsPerPage={pagination.perPage}
-					caption={tableCaption}
-					sort={sort}
-					setSort={setSort}
-					totalItems={totalItems}
-					ref={tableRef}
+					<SortBySelect sort={sort} setSort={setSort} />
+				</FilterBar>
+				<FilterAccordion
+					id={bodyId}
+					ariaLabelledBy={buttonId}
+					filters={filters}
+					isOpen={isOpen}
+					setFilters={setFilters}
+					resetFilters={resetFilters}
 				/>
+				<ActiveFilters
+					filters={filters}
+					removeFilter={removeFilter}
+					resetFilters={resetFilters}
+				/>
+			</FilterRegion>
 
-				{data.length ? (
-					<Stack>
-						<PaginationButtons
-							currentPage={pagination.page}
-							onChange={(page) => setPagination({ ...pagination, page })}
-							totalPages={totalPages}
-						/>
-					</Stack>
-				) : null}
-			</Stack>
-		</PageContent>
+			<DashboardTable
+				data={data}
+				loading={loading}
+				itemsPerPage={pagination.perPage}
+				caption={tableCaption}
+				sort={sort}
+				setSort={setSort}
+				totalItems={totalItems}
+				ref={tableRef}
+			/>
+
+			{data.length ? (
+				<Stack>
+					<PaginationButtons
+						currentPage={pagination.page}
+						onChange={(page) => setPagination({ ...pagination, page })}
+						totalPages={totalPages}
+					/>
+				</Stack>
+			) : null}
+		</Stack>
 	);
 };
